+++ Test.py
@@ -27,4 +27,4 @@ model=bc.train(df=df,target=target,features=features)
 model.stats()
 #print(model.predict([[1,1,1,1,1]]))
 #print(model.features())
-#model.save()
+model.save()
